MouseEvent copy constructor fix, and minor viewmanager type fix.

+++ modules/basegl/viewmanager.cpp
@@ -41,7 +41,7 @@ Event* ViewManager::registerEvent(Event* event) {
         activePosition_ = flipY(mouseEvent->pos(), mouseEvent->canvasSize());
         if (!viewportActive_ && mouseEvent->state() == MouseEvent::MOUSE_STATE_PRESS) {
             viewportActive_ = true;
-            activeView_ = static_cast<int>(findView(activePosition_));
+            activeView_ = findView(activePosition_);
         } else if (viewportActive_ && mouseEvent->state() == MouseEvent::MOUSE_STATE_RELEASE) {
             viewportActive_ = false;
         }
@@ -61,7 +61,7 @@ Event* ViewManager::registerEvent(Event* event) {
                                 gestureEvent->canvasSize());
         if (!viewportActive_ && gestureEvent->state() == GestureEvent::GESTURE_STATE_STARTED) {
             viewportActive_ = true;
-            activeView_ = static_cast<int>(findView(activePosition_));
+            activeView_ = findView(activePosition_);
         } else if (viewportActive_ && gestureEvent->state() == GestureEvent::GESTURE_STATE_ENDED) {
             viewportActive_ = false;
         }
@@ -82,7 +82,7 @@ Event* ViewManager::registerEvent(Event* event) {
         activePosition_ = flipY(touchEvent->getCenterPoint(), touchEvent->canvasSize());
         if (!viewportActive_ && touchEvent->getTouchPoints().front().state() == TouchPoint::TOUCH_STATE_STARTED) {
             viewportActive_ = true;
-            activeView_ = static_cast<int>(findView(activePosition_));
+            activeView_ = findView(activePosition_);
         } else if (viewportActive_ && touchEvent->getTouchPoints().front().state() ==
             TouchPoint::TOUCH_STATE_ENDED) {
             viewportActive_ = false;
@@ -125,8 +125,8 @@ size_t ViewManager::size() const { return views_.size(); }
 
 void ViewManager::clear() { views_.clear(); }
 
-size_t ViewManager::findView(ivec2 pos) const {
-    for (size_t i = 0; i < views_.size(); ++i) {
+int ViewManager::findView(ivec2 pos) const {
+    for (int i = 0; i < static_cast<int>(views_.size()); ++i) {
         if (inView(views_[i], pos)) {
             return i;
         }
